+++ packages/app-check/ios/RNFBAppCheck/RNFBAppCheckProviderFactory.h
@@ -15,7 +15,7 @@
  *
  */
 
-#import <FirebaseAppCheck/FIRAppCheck.h>
+#import "RNFBAppCheckProvider.h"
 
 @interface RNFBAppCheckProviderFactory : NSObject <FIRAppCheckProviderFactory>