+++ packages/app/android/src/reactnative/java/io/invertase/firebase/common/RCTConvertFirebase.java
@@ -98,7 +98,7 @@ public class RCTConvertFirebase {
 
     if (appConfig.hasKey("automaticDataCollectionEnabled")) {
       firebaseApp.setDataCollectionDefaultEnabled(
-          appConfig.getBoolean("automaticDataCollectionEnabled"));
+          Boolean.valueOf(appConfig.getBoolean("automaticDataCollectionEnabled")));
     }
 
     if (appConfig.hasKey("automaticResourceManagement")) {