+++ packages/auth/ios/RNFBAuth/RNFBAuthModule.m
@@ -55,8 +55,10 @@ static __strong NSMutableDictionary *idTokenHandlers;
 static __strong NSMutableDictionary *emulatorConfigs;
 // Used for caching credentials between method calls.
 static __strong NSMutableDictionary<NSString *, FIRAuthCredential *> *credentials;
+#if TARGET_OS_IOS
 static __strong NSMutableDictionary<NSString *, FIRMultiFactorResolver *> *cachedResolver;
 static __strong NSMutableDictionary<NSString *, FIRMultiFactorSession *> *cachedSessions;
+#endif
 
 @implementation RNFBAuthModule
 #pragma mark -
@@ -76,8 +78,10 @@ RCT_EXPORT_MODULE();
     idTokenHandlers = [[NSMutableDictionary alloc] init];
     emulatorConfigs = [[NSMutableDictionary alloc] init];
     credentials = [[NSMutableDictionary alloc] init];
+#if TARGET_OS_IOS
     cachedResolver = [[NSMutableDictionary alloc] init];
     cachedSessions = [[NSMutableDictionary alloc] init];
+#endif
   });
   return self;
 }
@@ -103,8 +107,10 @@ RCT_EXPORT_MODULE();
   [idTokenHandlers removeAllObjects];
 
   [credentials removeAllObjects];
+#if TARGET_OS_IOS
   [cachedResolver removeAllObjects];
   [cachedSessions removeAllObjects];
+#endif
 }
 
 #pragma mark -
@@ -415,6 +421,7 @@ RCT_EXPORT_METHOD(updatePassword
   }
 }
 
+#if TARGET_OS_IOS
 RCT_EXPORT_METHOD(updatePhoneNumber
                   : (FIRApp *)firebaseApp
                   : (NSString *)provider
@@ -455,6 +462,7 @@ RCT_EXPORT_METHOD(updatePhoneNumber
     [self promiseNoUser:resolve rejecter:reject isError:YES];
   }
 }
+#endif
 
 RCT_EXPORT_METHOD(updateProfile
                   : (FIRApp *)firebaseApp
@@ -617,6 +625,7 @@ RCT_EXPORT_METHOD(signInWithProvider
     [builder setCustomParameters:provider[@"customParameters"]];
   }
 
+#if TARGET_OS_IOS
   [builder getCredentialWithUIDelegate:nil
                             completion:^(FIRAuthCredential *_Nullable credential,
                                          NSError *_Nullable error) {
@@ -647,6 +656,7 @@ RCT_EXPORT_METHOD(signInWithProvider
                                               }];
                               }
                             }];
+#endif
 }
 
 RCT_EXPORT_METHOD(confirmPasswordReset
@@ -817,6 +827,7 @@ RCT_EXPORT_METHOD(signInWithCustomToken
                  }];
 }
 
+#if TARGET_OS_IOS
 RCT_EXPORT_METHOD(signInWithPhoneNumber
                   : (FIRApp *)firebaseApp
                   : (NSString *)phoneNumber
@@ -1053,6 +1064,7 @@ RCT_EXPORT_METHOD(confirmationResultConfirm
                   }
                 }];
 }
+#endif
 
 RCT_EXPORT_METHOD(linkWithCredential
                   : (FIRApp *)firebaseApp
@@ -1122,6 +1134,7 @@ RCT_EXPORT_METHOD(linkWithProvider
     [builder setCustomParameters:provider[@"parameters"]];
   }
 
+#if TARGET_OS_IOS
   [builder getCredentialWithUIDelegate:nil
                             completion:^(FIRAuthCredential *_Nullable credential,
                                          NSError *_Nullable error) {
@@ -1151,6 +1164,7 @@ RCT_EXPORT_METHOD(linkWithProvider
                                               }];
                               }
                             }];
+#endif
 }
 
 RCT_EXPORT_METHOD(unlink
@@ -1248,7 +1262,7 @@ RCT_EXPORT_METHOD(reauthenticateWithProvider
   if (provider[@"parameters"]) {
     [builder setCustomParameters:provider[@"parameters"]];
   }
-
+#if TARGET_OS_IOS
   [builder getCredentialWithUIDelegate:nil
                             completion:^(FIRAuthCredential *_Nullable credential,
                                          NSError *_Nullable error) {
@@ -1279,6 +1293,7 @@ RCT_EXPORT_METHOD(reauthenticateWithProvider
                                                         }];
                               }
                             }];
+#endif
 }
 
 RCT_EXPORT_METHOD(fetchSignInMethodsForEmail
@@ -1375,10 +1390,12 @@ RCT_EXPORT_METHOD(useEmulator
   } else if ([provider compare:@"github.com" options:NSCaseInsensitiveSearch] == NSOrderedSame) {
     credential = [FIRGitHubAuthProvider credentialWithToken:authToken];
   } else if ([provider compare:@"phone" options:NSCaseInsensitiveSearch] == NSOrderedSame) {
+#if TARGET_OS_IOS
     DLog(@"using app credGen: %@", firebaseApp.name) credential =
         [[FIRPhoneAuthProvider providerWithAuth:[FIRAuth authWithApp:firebaseApp]]
             credentialWithVerificationID:authToken
                         verificationCode:authTokenSecret];
+#endif
   } else if ([provider compare:@"oauth" options:NSCaseInsensitiveSearch] == NSOrderedSame) {
     credential = [FIROAuthProvider credentialWithProviderID:@"oauth"
                                                     IDToken:authToken
@@ -1423,6 +1440,7 @@ RCT_EXPORT_METHOD(useEmulator
   }
 }
 
+#if TARGET_OS_IOS
 - (NSDictionary *)multiFactorResolverToDict:(FIRMultiFactorResolver *)resolver {
   // Temporarily store the non-serializable session for later
   NSString *sessionHash = [NSString stringWithFormat:@"%@", @([resolver.session hash])];
@@ -1433,6 +1451,7 @@ RCT_EXPORT_METHOD(useEmulator
     @"auth" : resolver.auth
   };
 }
+#endif
 
 - (NSString *)getJSFactorId:(NSString *)factorId {
   if ([factorId isEqualToString:@"1"]) {
@@ -1543,6 +1562,7 @@ RCT_EXPORT_METHOD(useEmulator
   }
 
   NSDictionary *resolverDict = nil;
+#if TARGET_OS_IOS
   if ([error userInfo][FIRAuthErrorUserInfoMultiFactorResolverKey] != nil) {
     FIRMultiFactorResolver *resolver = error.userInfo[FIRAuthErrorUserInfoMultiFactorResolverKey];
     resolverDict = [self multiFactorResolverToDict:resolver];
@@ -1550,6 +1570,7 @@ RCT_EXPORT_METHOD(useEmulator
     NSString *sessionKey = [NSString stringWithFormat:@"%@", @([resolver.session hash])];
     cachedResolver[sessionKey] = resolver;
   }
+#endif
 
   return @{
     @"code" : code,
@@ -1696,11 +1717,14 @@ RCT_EXPORT_METHOD(useEmulator
     @"refreshToken" : user.refreshToken,
     @"tenantId" : user.tenantID ? (id)user.tenantID : [NSNull null],
     keyUid : user.uid,
+#if TARGET_OS_IOS
     @"multiFactor" :
         @{@"enrolledFactors" : [self convertMultiFactorData:user.multiFactor.enrolledFactors]}
+#endif
   };
 }
 
+#if TARGET_OS_IOS
 - (NSArray<NSMutableDictionary *> *)convertMultiFactorData:(NSArray<FIRMultiFactorInfo *> *)hints {
   NSMutableArray *enrolledFactors = [NSMutableArray array];
 
@@ -1720,6 +1744,7 @@ RCT_EXPORT_METHOD(useEmulator
   }
   return enrolledFactors;
 }
+#endif
 
 - (NSDictionary *)authCredentialToDict:(FIRAuthCredential *)authCredential {
   NSString *authCredentialHash = [NSString stringWithFormat:@"%@", @([authCredential hash])];